Backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
setuptools==69.5.1
sqlparse==0.5.0
swagger-spec-validator==3.0.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
